relocating IdLookupMixin from common/api/mixins.py to common/serializers.py
as per code review notes, I looked into replacing for example:

study_population = StudyPopulationSerializer() # where StudyPopulationSerializer mixes in IdLookupMixin

with:
study_population = serializers.PrimaryKeyRelatedField(queryset=models.StudyPopulation.objects)

This works, I can pass in the number for the studypop into e.g. a comparison set upload. But, if I do
that then all I get back out in the returned JSON is the id, not the full object. And
various existing tests/exports assume that the full JSON is going to be there. So leaving this intact and
just relocating IdLookupMixin into common/serializers as suggested.
